Move function to count open files to fastos. Count open files on Darwin.
